Added tech_news_concern to allow use elsewhere
